illumina2cluster/prep_sample_sheet.py: refactored into bcftbx.cli and 'bin' directory.
