+++ bin/prep_sample_sheet.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python
+#
+#     prep_sample_sheet.py: prepare sample sheet file for Illumina sequencers
+#     Copyright (C) University of Manchester 2021 Peter Briggs
+#
+from bcftbx.cli.prep_sample_sheet import main
+if __name__ == "__main__":
+     main()
